refactor(mobile): split ai chat store
- extract store state helper
- add store state tests
